+++ src/pages/mypage/_components/ProductList.tsx
@@ -1,12 +1,13 @@
+import ProductCardList from '@/components/feature/product/ProductCardList/ProductCardList';
+import { Spinner } from '@/components/shared';
+import useGetMe from '@/models/queries/auth/useGetMe';
+import useGetCreatedProducts from '@/models/queries/user/products/created-products/useGetCreatedProducts';
+import useGetFavoriteProducts from '@/models/queries/user/products/favorite-products/useGetFavoriteProducts';
+import useGetReviewedProducts from '@/models/queries/user/products/reviewed-products/useGetReviewedProducts';
+import { ProductCategory } from '@/pages/user/[userId]';
+
 import { useEffect, useMemo, useCallback } from 'react';
 import { useInView } from 'react-intersection-observer';
-import ProductCardList from '@/shared/components/ProductCardList/ProductCardList';
-import Spinner from '@/shared/components/Spinner/Spinner';
-import useGetCreatedProducts from '@/models/user/products/created-products/useGetCreatedProducts';
-import useGetFavoriteProducts from '@/models/user/products/favorite-products/useGetFavoriteProducts';
-import useGetReviewedProducts from '@/models/user/products/reviewed-products/useGetReviewedProducts';
-import useGetMe from '@/models/auth/useGetMe';
-import { ProductCategory } from '..';
 
 interface ProductListProps {
   selectedCategory: ProductCategory;
