+++ src/pages/user/[userId]/_components/ProductList.tsx
@@ -1,11 +1,13 @@
+import { useParams } from 'next/navigation';
 import { useEffect, useMemo, useCallback } from 'react';
 import { useInView } from 'react-intersection-observer';
-import ProductCardList from '@/shared/components/ProductCardList/ProductCardList';
-import Spinner from '@/shared/components/Spinner/Spinner';
-import useGetCreatedProducts from '@/models/user/products/created-products/useGetCreatedProducts';
-import useGetFavoriteProducts from '@/models/user/products/favorite-products/useGetFavoriteProducts';
-import useGetReviewedProducts from '@/models/user/products/reviewed-products/useGetReviewedProducts';
-import { useParams } from 'next/navigation';
+
+import useGetCreatedProducts from '@/models/queries/user/products/created-products/useGetCreatedProducts';
+import useGetFavoriteProducts from '@/models/queries/user/products/favorite-products/useGetFavoriteProducts';
+import useGetReviewedProducts from '@/models/queries/user/products/reviewed-products/useGetReviewedProducts';
+
+import { Spinner } from '@/components/shared';
+import ProductCardList from '@/components/feature/product/ProductCardList/ProductCardList';
 import { ProductCategory } from '..';
 
 interface ProductListProps {
