+++ src/pages/user/[userId]/index.tsx
@@ -1,19 +1,20 @@
 import { useState } from 'react';
 import { useParams } from 'next/navigation';
-import { ChevronDown } from 'lucide-react';
-import { useInView } from 'react-intersection-observer';
 
 import useGetMe from '@/models/queries/auth/useGetMe';
-import useGetCreatedProducts from '@/models/queries/user/products/created-products/useGetCreatedProducts';
-import useGetFavoriteProducts from '@/models/queries/user/products/favorite-products/useGetFavoriteProducts';
-import useGetReviewedProducts from '@/models/queries/user/products/reviewed-products/useGetReviewedProducts';
 import useUserProfile from '@/models/queries/user/profile/useUserProfile';
 
 import MogazoaLayout from '@/components/layout/App/MogazoaLayout';
-import ActivityCard from '@/components/feature/product/ActivityCard/ActivityCard';
-import ProductCardList from '@/components/feature/product/ProductCardList/ProductCardList';
 import ProfileCard from '@/components/feature/profile/ProfileCard/ProfileCard';
-import { Floating, Spinner } from '@/components/shared';
+import ActivitySection from './_components/ActivitySection';
+import ProductCategorySelector from './_components/ProductCategorySelector';
+import ProductList from './_components/ProductList';
+
+export enum ProductCategory {
+  REVIEWED = '리뷰 남긴 상품',
+  CREATED = '등록한 상품',
+  FAVORITE = '찜한 상품',
+}
 
 const UserProfile = () => {
   const params = useParams();
